cleanup prover
