+++ plonk/code/src/protocol/arithmetizer/wire/mod.rs
@@ -7,7 +7,7 @@ mod op_u64;
 mod op_usize;
 mod op_wire;
 
-use crate::util::{if_debug, is_debug};
+use crate::util::misc::{if_debug, is_debug};
 
 use super::{Arithmetizer, WireID};
 use ast::WireAST;
@@ -76,7 +76,7 @@ impl fmt::Display for Wire {
 
 #[cfg(test)]
 mod tests {
-    use crate::util::map_to_alphabet;
+    use crate::util::misc::map_to_alphabet;
 
     use super::Arithmetizer;
     use super::*;
