+++ plonk/code/src/protocol/plonk/mod.rs
@@ -1,6 +1,9 @@
 mod instance;
+mod pi;
 mod proof;
 mod transcript;
+mod verify;
 
+pub use pi::{EvalProofs, Proof, ProofCommitments, ProofEvaluations};
 pub use proof::prove;
-pub use proof::verify;
+pub use verify::verify;
